# msg_helpers: improve perf of unpackString by 20% and halve allocs