--- a/msg_helpers.go
+++ b/msg_helpers.go
@@ -261,28 +261,42 @@ func packUint64(i uint64, msg []byte, off int) (off1 int, err error) {
 }
 
 func unpackString(msg []byte, off int) (string, int, error) {
-	if off+1 > len(msg) {
+	start := off + 1
+	if start > len(msg) {
 		return "", off, &Error{err: "overflow unpacking txt"}
 	}
 	l := int(msg[off])
-	if off+l+1 > len(msg) {
+	end := start + l
+	if end > len(msg) {
 		return "", off, &Error{err: "overflow unpacking txt"}
 	}
 	var s strings.Builder
-	s.Grow(l)
-	for _, b := range msg[off+1 : off+1+l] {
+	p := msg[start:end]
+	for i, b := range p {
 		switch {
 		case b == '"' || b == '\\':
+			if s.Len() == 0 {
+				s.Grow(l * 2)
+				s.Write(p[:i])
+			}
 			s.WriteByte('\\')
 			s.WriteByte(b)
 		case b < ' ' || b > '~': // unprintable
+			if s.Len() == 0 {
+				s.Grow(l * 2)
+				s.Write(p[:i])
+			}
 			s.WriteString(escapeByte(b))
 		default:
-			s.WriteByte(b)
+			if s.Len() != 0 {
+				s.WriteByte(b)
+			}
 		}
 	}
-	off += 1 + l
-	return s.String(), off, nil
+	if s.Len() == 0 {
+		return string(p), end, nil
+	}
+	return s.String(), end, nil
 }
 
 func packString(s string, msg []byte, off int) (int, error) {