--- a/msg_helpers_test.go
+++ b/msg_helpers_test.go
@@ -136,3 +136,19 @@ func BenchmarkUnpackString(b *testing.B) {
 		}
 	}
 }
+
+func BenchmarkUnpackString_NoEscape(b *testing.B) {
+	msg := []byte("\x00large.example.com")
+	msg[0] = byte(len(msg) - 1)
+
+	for n := 0; n < b.N; n++ {
+		got, _, err := unpackString(msg, 0)
+		if err != nil {
+			b.Fatal(err)
+		}
+
+		if want := `large.example.com`; want != got {
+			b.Errorf("expected %q, got %q", want, got)
+		}
+	}
+}