**Edit:** forgot to include results from: `BenchmarkUnpackString_NoEscape`
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkUnpackString-12              79.7          64.5          -19.07%
BenchmarkUnpackString_NoEscape-12     51.3          51.8          +0.97%

benchmark                             old allocs     new allocs     delta
BenchmarkUnpackString-12              2              1              -50.00%
BenchmarkUnpackString_NoEscape-12     1              1              +0.00%

benchmark                             old bytes     new bytes     delta
BenchmarkUnpackString-12              48            32            -33.33%
BenchmarkUnpackString_NoEscape-12     32            32            +0.00%
```