@tmthrgd 2019-09-21
I’m against this change as is. It adds a lot of complexity for relatively little gain (should only be observable when the string needs escaping).

That being said, I have an idea for how to do this cleaner that I’ll try out and send a PR for.