@charlievieth 2019-09-21
> I’m against this change as is. It adds a lot of complexity for relatively little gain (should only be observable when the string needs escaping).
> 
> That being said, I have an idea for how to do this cleaner that I’ll try out and send a PR for.

The only complexity here is the lazy allocation of the `strings.Builder`, which boils down to a few `if` statements.  Remember the real gain here is the reduction of allocs, which reduces GC pressure and is critical to improving the throughput of things like CoreDNS (which is how I identified this function).